Move database manager factory class
